requirements.txt
flask
python-bitcoinlib
bip32utils
pbkdf2
requests
pymongo